Added user verification by role and error handling

+++ src/activities/activities.controller.ts
@@ -1,33 +1,43 @@
 // activities.controller.ts
-import { Controller, Post, Body, Get, Param, Put, Delete, Req, UseGuards, NotFoundException } from '@nestjs/common';
+import { Controller, Post, Body, Get, Param, Put, Delete, Req, UseGuards, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { Activity } from './activities.schema';
 import { ActivitiesService } from './activities.service';
-import { Auth0Guard } from 'src/auth/auth0.guard';
-import { Response, Request } from 'express';
+import { Auth0Guard, auth0Access } from 'src/auth/auth0.guard';
+import { Request } from 'express';
 import axios from 'axios';
 
 @Controller('rest/activities')
 @UseGuards(Auth0Guard)
 export class ActivitiesController {
   constructor(private readonly activitiesService: ActivitiesService) {}
-  
+
   private extractTokenFromHeader(request: Request): string | undefined {
     const [type, token] = request.headers.authorization?.split(' ') ?? [];
     return type === 'Bearer' ? token : undefined;
   }
 
-  // Función para obtener información del usuario desde el token de autorización
   private async getUserFromToken(token: string): Promise<any> {
     try {
       const response = await axios.get('https://fct-netex.eu.auth0.com/userinfo', {
-
         headers: {
           Authorization: `Bearer ${token}`,
         },
       });
       return response.data;
     } catch (error) {
-      throw new NotFoundException('User not found');
+      throw new NotFoundException('Usuario no encontrado');
+    }
+  }
+
+  private async verifyUserIsAdmin(userSub: string): Promise<void> {
+    const userRolesResponse = await axios.get(`https://fct-netex.eu.auth0.com/api/v2/users/${encodeURIComponent(userSub)}/roles`, {
+      headers: {
+        Authorization: `Bearer ${await auth0Access()}`,
+      },
+    });
+    const userRoles = userRolesResponse.data;
+    if (!userRoles.some(role => role.name === "admin")) {
+      throw new UnauthorizedException('Usuario no autorizado');
     }
   }
 
@@ -35,12 +45,13 @@ export class ActivitiesController {
   async createActivity(@Req() request: Request, @Body() activityData: Partial<Activity>): Promise<Activity> {
     const token = this.extractTokenFromHeader(request);
     if (!token) {
-      throw new NotFoundException('Token not found');
+      throw new NotFoundException('Token no encontrado');
     }
     const user = await this.getUserFromToken(token);
+    await this.verifyUserIsAdmin(user.sub);
     return this.activitiesService.createActivity(activityData, user.sub);
   }
-  
+
   @Get()
   async getAllActivities(): Promise<Activity[]> {
     return this.activitiesService.getAllActivities();
@@ -52,15 +63,24 @@ export class ActivitiesController {
   }
 
   @Put(':id')
-  async updateActivity(
-    @Param('id') id: string,
-    @Body() updatedActivityData: Partial<Activity>,
-  ): Promise<Activity> {
-    return this.activitiesService.updateActivity(id, updatedActivityData);
+  async updateActivity(@Param('id') id: string, @Body() activityData: Partial<Activity>, @Req() request: Request): Promise<Activity> {
+    const token = this.extractTokenFromHeader(request);
+    if (!token) {
+      throw new NotFoundException('Token no encontrado');
+    }
+    const user = await this.getUserFromToken(token);
+    await this.verifyUserIsAdmin(user.sub);
+    return this.activitiesService.updateActivity(id, activityData);
   }
 
   @Delete(':id')
-  async deleteActivity(@Param('id') id: string): Promise<Activity> {
+  async deleteActivity(@Param('id') id: string, @Req() request: Request): Promise<Activity> {
+    const token = this.extractTokenFromHeader(request);
+    if (!token) {
+      throw new NotFoundException('Token no encontrado');
+    }
+    const user = await this.getUserFromToken(token);
+    await this.verifyUserIsAdmin(user.sub);
     return this.activitiesService.deleteActivity(id);
   }
 }
